refactor(deck): introduce mode-switching deck view
Replace monolithic deck-view.vue with a mode-driven shell that picks
between CardGrid and CardEditor. Normalize view roots to index.vue
(dashboard, deck), rename overview-panel.vue to deck-hero.vue, drop
unused bulk-select-toolbar.vue, and remove the shop route.
