+++ src/views/deck/deck-hero.vue
